lambda-packaging: add suport for INTERMEDIATE_PACKAGING variable
If INTERMEDIATE_PACKAGING is set to true for `package_lambda_runner.sh`
it's assumed that the output of a build stage is meant as an input to
the next one in the pipeline and thus is output to `work` directory.

If it's set to false, the resulting zip will be output to `result`
directory.

The main use-case for this is providing one package's build output into
another package (e.g. as is the case with `common`).